refactor: svg 파일 사이즈 삭제.
